Create an Options struct, just with a layout field for now
We'll move all the global variables for command-line options into a
struct.  This will make it easier to isolate the code that deals with
the command line, and probably to write tests for different
combinations of options.

General strategy:

* options.h has the Options struct.

* optionsInit() sets the default values.

* For now, main() still parses the options from argc/argv.  We'll
probably do it in a separate function at some later point.