Create an ImageProcessParameters struct for the constant data in the pipeline
Start by moving these globals into that struct:

  deskewScanRangeRad
  deskewScanStepRad
  deskewScanDeviationRad

This might seem inconsequential for now, but it lets us start a place
to put the data that is computed after obtaining the command-line
options, but that stays constant throughout the rest of the program.
This is the "configuration" for the image processing pipeline, if you
will.